build-and-test.sh: Extratime 1s, do not run compare on RE and TLE
Extratime was 2*TL+5, it is now 2*TL+1 which is more sane when running
all tests.

Also it will not run the compare script when RE and TLE, the output of
these results contains pretty much all the output file and takes DIFF
(and the compare script) a long time to run.

Signed-off-by: Bruno Ribas <brunoribas@gmail.com>