calibreitor.sh: Add Option to Allow TLE during Calibration
ALLOWTLEDURINGCALIBRATION=y

Allow the good solution to take time limit during calibration. This
allows for the addition of new types of problems to the system,
problems that the good solution may not necessarily be able to solve
for all test cases. For example, planning and optimization problems.

Remember to adjust the TLMOD[calibrafactor] variable correctly.