build-and-test.sh: Introduce configurable variable TLERERUN
The TLERERUN variable is used to control the behavior of tests that
receive a Time Limit Exceeded (TLE) verdict. By default, the value is
set to "y", which means that a test that receives a TLE verdict will
be run one more time to ensure accuracy. If the problem allows
parallel tests, this additional test will be performed individually on
a separate CPU.

However, certain problems, such as those that do not support parallel
testing, may disable this secondary test. This is particularly useful
for problems with a high Time Limit and where retesting after a TLE
verdict is not necessary in some cases.

Signed-off-by: Bruno Ribas <brunoribas@gmail.com>