build-and-test.sh: Run COMPARE for every run early
Now build-and-test.sh will run the COMPARE script just after the run
has been executed, this will allow for ealier decisions regarding
failed tests.

Also it makes a simpler way to define which Verdict is the worse, so
the final verdict is the worse found during execution. This rank can
be done via problem parameter later.

Signed-off-by: Bruno Ribas <brunoribas@gmail.com>