build-and-test.sh: Handle return values >= 127 as RuntimeError
This modification changes the interpretation of process exit codes in
`build-and-test.sh` to treat any return value of 127 or greater as a
RuntimeError (RE). This adjustment aligns with Unix shell conventions,
which interpret exit codes in this range as indicating a process
termination due to signals, with the signal number being the exit code
subtracted by 128.

Specifically, an exit code of 127 is conventionally used to signify
"command not found", suggesting an attempt to execute a non-existent
file. By treating these codes as runtime errors, we improve the
script's ability to handle and report unusual termination conditions
in a more standardized manner.

Signed-off-by: Bruno Ribas <brunoribas@gmail.com>