build-and-test.sh: Add support for STOPWHEN_{TLE,WA,RE} variables
The default behavior of `build-and-test.sh`, especially when utilized
by `moj`, is to execute all tests for a given problem. This approach
ensures a detailed verdict on the success rate of submissions by
evaluating the percentage of correct answers.

However, this exhaustive testing model proves inefficient for problems
that are time-consuming to correct, notably those with extensive test
suites like those encountered in the FLIA course, where testing all
inputs might exceed 2 hours.

To accommodate the unique requirements of such problems — which accept
RE and TLE without penalization but mark WA with zero score regardless
of the number of correct instances — this update introduces the
`STOPWHEN_{TLE,WA,RE}` environment variables. These variables allow
for early termination of test execution based on the specified
criteria (TLE, WA, RE), optimizing the queue management by halting the
process for problems that are likely to fail, thereby saving time and
resources.

Key Enhancements:
- Add support for `STOPWHEN_{TLE,WA,RE}` environment variables in
`build-and-test.sh`.

- Facilitate early test termination for problems with lengthy
correction times, improving resource utilization.

- Address the specific grading policy of FLIA course problems,
enhancing the script's applicability and flexibility.

This modification streamlines the workflow for handling complex
problems, particularly when multiple incorrect solutions are present
in the queue, by preemptively stopping the testing process upon
encountering a WA verdict.

Signed-off-by: Bruno Ribas <brunoribas@gmail.com>